refactor(handler): refactor handler file structure
